+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.2.0 (27-06-2019)
+- added ability to handle empty arrays in query parameters
+
 ## 0.1.8 (27-06-2019)
 - bump swagger ui to 3.23.1